feat(Rust): Support automatic conversion between T and Option<T> when deserialize (#2563)
## Why?
Fory needs to support the following conversions when `deserialize`
(Option means nullable type)(left is sender_field_type, right is
receiver_field_type):
- value -> value
- value -> Option(value)
- Option(value) -> value
- Option(value) -> Option(value)
- Option(None) -> Option(None)
- Option(None) -> value_default

## What does this PR do?
- Support automatic conversion between `T` and `Option<T>` when
deserialize. The usage can be seen in
`test_complex_struct.rs::nullable()`.
- Add NullableFieldType to represent `T{nullable}`, while the original
FieldType exactly matches the data structure of a struct, such as
`Option{T}`. Comparison of `NullableFieldType` ignores the `nullable`
attribute.
- Add `read_var_uint32()` && `write_var_uint32()`.
- Changed the type of `type_id` from `i16` to `u32`, and updated the
read/write operations from `i16()` to `var_uint32()`
- When registering a struct, add `types::STRUCT` to its id in addition
to the original type_id
- Removed the sending of `meta_index`.
- For reuse, extract the `read_compatible` function.



## Detail
1. The processing approach is as follows:
- At compile time, parse and obtain the `FieldType` instance. The
`FieldType` (e.g., `Option{T}`) can be converted into a
`NullableFieldType` (e.g., `T{nullable=true}`).
- At compile time, based on the `NullableFieldType`, generate a
`_deserialize_nullable_{field_name}` function for each field. This
function defines the read process constrained by the
`local_nullable_field_type` (compile-time instance, obtained through
macro parsing) and determines the actual data read according to the
`remote_nullable_field_type` (runtime instance, obtained from the
received type_meta).
   - Deserialization flow:
     ```rust
     if local_field_name_f1 == remote_field_name {
         if local_field_type_f1 == remote_field_type {
             f1 = <f1_T>::deserialize();
         } else {
if nullable_local_field_type_f1 == nullable_remote_field_type {
f1 = Self::_deserialize_nullable_f1(remote_nullable_field_type);
             } else {
                 skip_field_value(remote_field_type);
                 f1 = default();
             }
         }
     } else if local_field_name_f2 == remote_field_name {
         ...
     } else if ... {
         ...
     } else {
         skip_field_value(remote_field_type);
     }
     ```
2. Adjacent `Option`s like `Option<Option<T>>` are not supported. If
such a field is defined, a compile-time error will be raised: `adjacent
Options are not supported`.

## Related issues
#2529 

## Does this PR introduce any user-facing change?

- [ ] Does this PR introduce any public API change?
- [x] Does this PR introduce any binary protocol compatibility change?

## Benchmark

+++ rust/fory-core/src/resolver/type_resolver.rs
@@ -48,18 +48,18 @@ impl Harness {
 
 pub struct TypeInfo {
     type_def: Vec<u8>,
-    type_id: i16,
+    type_id: u32,
 }
 
 impl TypeInfo {
-    pub fn new<T: StructSerializer>(fory: &Fory, type_id: i16) -> TypeInfo {
+    pub fn new<T: StructSerializer>(fory: &Fory, type_id: u32) -> TypeInfo {
         TypeInfo {
             type_def: T::type_def(fory, type_id),
             type_id,
         }
     }
 
-    pub fn get_type_id(&self) -> i16 {
+    pub fn get_type_id(&self) -> u32 {
         self.type_id
     }
 
@@ -70,8 +70,8 @@ impl TypeInfo {
 
 #[derive(Default)]
 pub struct TypeResolver {
-    serialize_map: HashMap<i16, Harness>,
-    type_id_map: HashMap<std::any::TypeId, i16>,
+    serialize_map: HashMap<u32, Harness>,
+    type_id_map: HashMap<std::any::TypeId, u32>,
     type_info_map: HashMap<std::any::TypeId, TypeInfo>,
 }
 
@@ -85,7 +85,7 @@ impl TypeResolver {
         })
     }
 
-    pub fn register<T: StructSerializer>(&mut self, type_info: TypeInfo, id: i16) {
+    pub fn register<T: StructSerializer>(&mut self, type_info: TypeInfo, id: u32) {
         fn serializer<T2: 'static + StructSerializer>(this: &dyn Any, context: &mut WriteContext) {
             let this = this.downcast_ref::<T2>();
             match this {
@@ -115,7 +115,7 @@ impl TypeResolver {
         self.get_harness(*self.type_id_map.get(&type_id).unwrap())
     }
 
-    pub fn get_harness(&self, id: i16) -> Option<&Harness> {
+    pub fn get_harness(&self, id: u32) -> Option<&Harness> {
         self.serialize_map.get(&id)
     }
 }
